Validate coverage JSON before writing it to file
This is primarily intended for early error detection, to catch unexpected
inputs (and/or bugs) that may lead to us producing invalid coverage files.

Throwing an error in Code Base Investigator is likely to be more helpful
than waiting until the user tries to load the coverage file into some other
tool.

Signed-off-by: John Pennycook <john.pennycook@intel.com>