Add support for automatic file discovery
When working with very large projects, defining the set of source files that
should be compiled (even via a regular expression) is non-trivial. It may also
be unnecessary effort, since the set of files may be described precisely by a
compilation database.

This commit adds special meaning to codebase configurations with empty file
lists (i.e., "files: []"), interpreted as a request to build the list of files
automatically from the platform configuration(s).

Explicit file lists that resolve to an empty list (e.g., because of invalid
regular expressions) continue to throw an error as before.

Closes #13.

Signed-off-by: John Pennycook <john.pennycook@intel.com>
Co-authored-by: Jason Sewall <jason.sewall@gmail.com>