Update exclude logic to support file discovery
Previously, files were only included in divergence analysis if they were
specified as part of the codebase or appeared in a compilation database.
Header files not specified in the codebase were always excluded.

This commit ensures that any additional files in the root directory that
are encountered during parsing are included in divergence analysis by default.

Signed-off-by: John Pennycook <john.pennycook@intel.com>
Co-authored-by: Jason Sewall <jason.sewall@gmail.com>