Account for common implicit compiler behavior
This commit introduces functionality to detect a compiler (based on its name)
and ensure that implicit behavior is reflected in the defines and includes.

This should be considered an initial and partial solution, which aims to
capture the most common sources of divergence when working with the following
compilers:

- clang
- icx, icpx and ifx
- nvcc

Placeholders exist for gcc and hipcc, but these do not yet enable any implicit
behavior.

Closes #15.

Signed-off-by: John Pennycook <john.pennycook@intel.com>