Replace file exclusion warning with info message
In real life scenarios, including a header (e.g., from the standard library)
will pull in a large number of additional headers. Printing a warning for
every system header encountered produces an overwhelming amount of output.

By moving this to an info message, users will not see this output unless they
opt-in to running in verbose mode.

Signed-off-by: John Pennycook <john.pennycook@intel.com>