Allow users to adjust compiler behavior
Although we can (and should) attempt to automatically respect the behavior of
various compilers, there are some behaviors that we cannot detect. For example,
many compilers define macros corresponding to the version of the compiler being
used, and it is impossible to derive this from the compiler name alone.

This commit introduces support for a per-codebase configuration file that can
be used to adjust the options used for specific compilers. This initial support
is limited to defining macros that should be defined in all passes. Eventually,
we should aim to build out this support until it is capable of describing all
of the compiler behaviors that are currently hard-coded into config.py.

Signed-off-by: John Pennycook <john.pennycook@intel.com>