Re-use exclude function across mapper and exporter
The definition of "exclude" was duplicated across mapper and exporter,
and so the two walkers disagreed on how to exclude a file from the
codebase.

By sharing the definition across the two classes, they are guaranteed to have
the same functionality, and it'll be harder for us to introduce these sorts of
bugs accidentally in the future.

Signed-off-by: John Pennycook <john.pennycook@intel.com>