Deprecate the codebasin package
The codebasin package was not originally designed for downstream users:
the interface is undocumented, and in many cases is not very user-friendly.

Semantic versioning recommends that there be at least one minor release with a
deprecation warning prior to removing functionality in a major release.
Deprecating the entire package is a simple solution that accurately
communicates our intent to users.

Signed-off-by: John Pennycook <john.pennycook@intel.com>