Simplify --platform (-p) option
After internal testing and review, we reached the conclusion that using -p to
both define and filter platforms was too confusing and difficult to teach.
Additionally, there are concerns that use of flags to define platforms may not
scale, since real-life use-cases may require more than a list of commands.

This commit is a partial revert of #49: the ability to filter platforms
remains, but the ability to define platforms directly via JSON is removed.

Signed-off-by: John Pennycook <john.pennycook@intel.com>