Update documentation to reflect CodeBase behavior
A side-effect of tracking all source files in the code base is that analysis
can now pick up unexpected files that were previously never encountered in a
compilation database.

For example, any C++ files automatically generated by CMake will be identified
as unused code if codebasin is invoked in a directory containing both build/
and src/ directories.

This commit updates the documentation to highlight the importance of running
codebasin in the right directory (or otherwise separating build and src).

Signed-off-by: John Pennycook <john.pennycook@intel.com>