Add progress bars to finder.find()
Parsing each file in the codebase and subsequent preprocessing are the
two most time-consuming steps of running codebasin.

Adding progress bars for these steps is expected to be a significant
improvement to user experience, as it enables a user to reason about how
fast codebasin is running and when it may complete.

Signed-off-by: John Pennycook <john.pennycook@intel.com>