Add ability to compute a "Code Utilization" score
This number provides an alternative view of a codebase that is complementary to
code divergence.

The code utilization is the average number of times each line of code is used:
the numerator is increased by 1 for each (SLOC, platform) combination, and the
denominator is increased by 1 for each SLOC.

The normalized utilization is the utilization divided by the number of
platforms to produce a number in the range [0, 1]: a score of 0 means that none
of the code is used by any of the platforms, while a score of 1 means that all
of the code is used by all of the platforms.

Signed-off-by: John Pennycook <john.pennycook@intel.com>