Fix definition and implementation of divergence
In the original paper, Harrell and Kitson define code divergence imprecisely in
terms of pairs (i, j) in a set. In our follow-up "Navigating P3" paper, we
attempted to make this more precise and in so doing introduced an error.
Specifically, we defined code divergence as drawing all pairs {i, j} from the
cartesian product H x H, which would have included {i, j} and {j, i}.

The computation of code divergence in Code Base Investigator was mostly
correct, but the incorrect notation was copied into the documentation.
The implementation also included a small bug, in that it evaluated the
divergence for the empty set and single-item sets as 0 instead of nan.

These issues were discovered while attempting to rewrite the divergence()
routine, and so this commit adds some regression tests to ensure that
divergence() produces the correct results.

Signed-off-by: John Pennycook <john.pennycook@intel.com>