Add files report
Introduces a new files report that visualizes the entire code base as a tree
with a format inspired by the "tree" utility.

Information about the platforms and SLOC used by each node in the tree is
printed similar to tree's "metafirst" output, in the format:

[Platforms | Used SLOC / Total SLOC | Code Divergence | Code Utilization]

Colors and formatting are all hard-coded in this initial version, but the
long-term plan is to allow configuration via the .cbi/config file.

Signed-off-by: John Pennycook <john.pennycook@intel.com>