Add instructions for virtual environments
Previously, we just linked to the documentation for "virtual environments".

Users have several options for creating these, and providing configuration
commands that can be copied-and-pasted into a terminal may improve the
out-of-box experience.

The changes in this commit add instructions for venv and uv. Each option is
displayed as a separate tab, so it should be straightforward to add more
options in the future.

Signed-off-by: John Pennycook <john.pennycook@intel.com>