Remove cbi.log from cbi-tree
The currently intended workflow for cbi-tree is to run codebasin first, and
ensure that it accurately analyzes the code base.  As such, we do not need to
display the full set of warnings from cbi-tree.

Over time, we can improve this workflow by making the relationship explicit
(e.g., by having codebasin produce a results directory that cbi-tree reads).

Signed-off-by: John Pennycook <john.pennycook@intel.com>