Add coverage workflow to GitHub Actions
This action performs a coverage analysis for any source files that have been
modified, and reports if there are any new or modified lines that are not
covered by any tests.

The coverage analysis is focused on new code because we know there are paths
that are currently not covered; the intent of this new action is only to
prevent a pull request from making the situation worse.

Signed-off-by: John Pennycook <john.pennycook@intel.com>