Support multiple flags in ExtendMatch
Previously, each "option string" defined its own set of passes.
This would produce the wrong behavior for any case where the
user defines an ExtendMatch action with multiple flags.

After this change, the flag name used as a key is the first
in the list of possible flags. This enables ExtendMatch actions
with multiple flags to store everything in a single list.

Signed-off-by: John Pennycook <john.pennycook@intel.com>