Rename Compiler and CompilerConfiguration classes
While sketching out documentation and other functionality, it became clear that
the names of these classes were pretty unclear. For example, an instance of the
"Compiler" class couldn't actually compile anything.

The new names are much clearer, and in keeping with what the classes do and how
they're used within CBI:

- An ArgumentParser parses arguments, and may be configured to emulate the
  behavior of specific compilers.

- A PreprocessorConfiguration describes the configuration of CBI's
  preprocessor.

Signed-off-by: John Pennycook <john.pennycook@intel.com>