Use f-string to format log.warning message
The % format in this warning message was broken, and printed "%s" instead of a
path. This may be something to do with our interception and coloring of warning
messages, but I can't be sure.

The simplest fix is to use an f-string (as we do for other warnings), because
that results in the string being formatted before it is passed to logging.

Signed-off-by: John Pennycook <john.pennycook@intel.com>