Recommend using cmake -G Ninja in the tutorial
It turns out that some generators (like Unix Makefiles) automatically compress
command-lines using response (.rsp) files if they exceed a certain number of
characters. There is no way to disable this behavior, and it is
language-specific.

Any options that are hidden inside an .rsp file is passed to the compiler using
a compiler-specific option, and the .rsp file itself may not exist when we try
to run our analysis.

Using Ninja avoids this issue, and thus produces better results.

Signed-off-by: John Pennycook <john.pennycook@intel.com>