Add progress tracker for identifying source files
If a source directory contains many files (e.g., because a build directory was
accidentally included) it can take a long time to identify the source files in
the code base. This commit adds a progress tracker with tqdm to make it clear
that CBI is doing something and hasn't just stalled.

We don't get a progress bar, because the number of files in the code base is
not known a priori, but tqdm will print the number of files found so far.

Signed-off-by: John Pennycook <john.pennycook@intel.com>